Added % System!
This is the server sided code that enables the possibility to remove a percentage (set in Config.lua if enabled!) of the persons money rather than a dedicated amount! This is ideal if you are looking for realistic roleplay, for example like in the UK how tax is 20%!

I have also beautified the updated code so that the code itself is not a mess! 😄